requirements.txt
appnope==0.1.2
argcomplete==1.12.3
attrs==21.4.0
Automat==20.2.0
backcall==0.2.0
boto3==1.20.47
botocore==1.23.49
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.1
constantly==15.1.0
cryptography==36.0.1
cssselect==1.1.0
cycler==0.11.0
decorator==5.1.0
dnspython==2.2.0
empyrical==0.5.5
grpcio==1.43.0
grpcio-status==1.43.0
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
ipython>=7.31.1
itemadapter==0.4.0
itemloaders==1.0.4
jedi==0.18.0
jellyfish==0.9.0
jmespath==0.10.0
joblib==1.1.0
kiwisolver==1.3.2
lxml==4.7.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
networkx==2.6.3
newsapi-python==0.2.6
nltk==3.6.7
numpy==1.21.4
packaging==21.3
pandas==1.3.4
pandas-datareader==0.10.0
parsel==1.6.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.0
pipx==0.16.4
priority==1.3.0
prompt-toolkit==3.0.22
Protego==0.1.16
proto-plus==1.19.8
protobuf==3.19.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
pyfolio==0.9.2
Pygments==2.10.0
pymongo==4.0.1
pyOpenSSL==21.0.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
queuelib==1.6.2
regex==2022.1.18
requests==2.27.1
rsa==4.8
s3transfer==0.5.1
scikit-learn==1.0.1
scipy==1.7.1
Scrapy==2.5.1
seaborn==0.11.2
segtok==1.5.11
service-identity==21.1.0
six==1.16.0
tabulate==0.8.9
threadpoolctl==3.0.0
tqdm==4.62.3
traitlets==5.1.1
Twisted==22.2.0
typing_extensions==4.0.1
urllib3==1.26.13
userpath==1.7.0
w3lib==1.22.0
wcwidth==0.2.5
yake==0.4.8
zope.interface==5.4.0
